Updates To 6.8.0, Fixes Missing Icons (#6)
It appears that there are assets missing beyond the finally listed icon
cases.

```swift
    case removeMediaImage = "remove-media-image"
    case removeMediaVideo = "remove-media-video"
    case removePage = "remove-page"
    case removePageAlt = "remove-page-alt"
```
**Acceptance Criteria** 
- All icons should be added, `A -> Z`

**Dev Note**
- Appears there is an issue with the `update.sh` where the only icons
downloaded are the ones that do not error out on github. In other words
there are too many for github to list and as a result only shows `N`
number of them then the remaining amount are not shown. Fetching and
downloading them w/ the script fails for that reason.

```sh
# Fetch icons from the GitHub repository and save them to the Assets.xcassets folder
icon_list=$(curl -sL "${GITHUB_REPO}" | grep -o 'blob/main/icons/[^"]*\.svg' | sed 's/blob\/main\//main\//')
```

**New Features!**
- - uiImage(from string:  String)
- - image(from string: String)

Adding the possibility to encode and decode more efficiently to and from
iconoir icons.