feat: Add new icons for alignment, app store, bridge surface, circle spark, constrained surface, curve array, number squares, post, and threads
This commit adds several new icons to the Iconoir enum. The added icons include alignHorizontalCenters, alignHorizontalSpacing, alignVerticalCenters, alignVerticalSpacing for alignment purposes. It also includes appStore for representing an app store icon.

Additionally, bridgeSurface and circleSpark are added as visual elements. For design purposes, constrainedSurface is included. CurveArray is added as a representation of a curved array shape. Number squares from 0 to 9 are introduced as number representations in square shapes.

Lastly, post and threads are included as symbols related to online discussions or forums.

Bug Fix
- Removes snapchat `fill="white"` property.